move all schema into one folder
