+++ gatekeeperschema/activate_user.go
@@ -1,0 +1,7 @@
+package gatekeeperschema
+
+// ActivateUserRequest ...
+type ActivateUserRequest struct {
+	Id          string `json:"id"`
+	NewPassword string `json:"new_password"`
+}
