+++ gatekeeperschema/change_password.go
@@ -1,0 +1,7 @@
+package gatekeeperschema
+
+// ChangePasswordRequest ...
+type ChangePasswordRequest struct {
+	Id          string `json:"id"`
+	NewPassword string `json:"new_password"`
+}
