+++ gatekeeperschema/confirm_password.go
@@ -1,0 +1,9 @@
+package gatekeeperschema
+
+// ConfirmPasswordRequest ...
+type ConfirmPasswordRequest struct {
+	Id              string `json:"id"`
+	OldPassword     string `json:"old_password"`
+	Password        string `json:"password"`
+	ConfirmPassword string `json:"confirm_password"`
+}
