+++ gatekeeperschema/organization_by_id.go
@@ -1,0 +1,6 @@
+package gatekeeperschema
+
+// GetOrganizationByIdRequest ...
+type GetOrganizationByIdRequest struct {
+	Id string `json:"id"`
+}
