+++ gatekeeperschema/soft_delete.go
@@ -1,0 +1,6 @@
+package gatekeeperschema
+
+// SoftDeleteRequest ...
+type SoftDeleteRequest struct {
+	Id string `json:"id"`
+}
