+++ gatekeeperschema/user_organization.go
@@ -1,0 +1,7 @@
+package gatekeeperschema
+
+// UpdateUserOrganizationRequest ...
+type UpdateUserOrganizationRequest struct {
+	UserIds        []string `json:"user_ids"`
+	OrganizationId string   `json:"organization_id"`
+}
