+++ gatekeeperschema/user_with_organization_by_id.go
@@ -1,0 +1,6 @@
+package gatekeeperschema
+
+// GetUserWithOrganizationByIdRequest ...
+type GetUserWithOrganizationByIdRequest struct {
+	Id string `json:"id"`
+}
