+++ gatekeeperschema/zendesk_user_id.go
@@ -1,0 +1,7 @@
+package gatekeeperschema
+
+// UpsertZendeskUserIdRequest ...
+type UpsertZendeskUserIdRequest struct {
+	ZendeskUserid string `json:"zendesk_userid"`
+	UserId        string `json:"user_id"`
+}
